# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DrRuby Web API",
    "version": "1.0.0",
    "description": "OpenAPI specification for DrRuby Web route handlers."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AppointmentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientUserId": {
            "type": "string"
          },
          "patientName": {
            "type": "string",
            "nullable": true
          },
          "staffUserId": {
            "type": "string",
            "nullable": true
          },
          "referralId": {
            "type": "string",
            "nullable": true
          },
          "scheduledAt": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/appointmentStatusSchema"
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "patientUserId",
          "patientName",
          "staffUserId",
          "referralId",
          "scheduledAt",
          "status",
          "note"
        ]
      },
...
      "appointmentStatusSchema": {
        "type": "string",
        "enum": [
          "scheduled",
          "done",
          "report_review",
          "consultation",
          "canceled"
        ]
      },
...
      "ScanRecordDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientUserId": {
            "type": "string"
          },
          "patientName": {
            "type": "string",
            "nullable": true
          },
          "objectKey": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string"
          },
          "inflammatoryScore": {
            "type": "number",
            "nullable": true
          },
          "pigmentationScore": {
            "type": "number",
            "nullable": true
          },
          "textureScore": {
            "type": "number",
            "nullable": true
          },
          "trend": {
            "$ref": "#/components/schemas/trendSchema",
            "nullable": true
          }
        },
        "required": [
          "id",
          "patientUserId",
          "patientName",
          "objectKey",
          "capturedAt",
          "inflammatoryScore",
          "pigmentationScore",
          "textureScore",
          "trend"
        ]
      },
...
      "trendSchema": {
        "type": "string",
        "enum": [
          "improving",
          "stable",
          "declining"
        ]
      },
      "UpdateAppointmentBody": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/appointmentStatusSchema"
              }
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UpdateScanRecordBody": {
        "type": "object",
        "properties": {
          "trend": {
            "$ref": "#/components/schemas/trendSchema",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/clinic/appointments/{id}": {
      "get": {
        "operationId": "get-clinic-appointments-{id}",
        "summary": "Get appointment detail",
        "description": "单条预约；归属不匹配按 404",
        "tags": [
          "Clinic"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "operationId": "patch-clinic-appointments-{id}",
        "summary": "Update appointment",
        "description": "推进状态 / 改备注",
        "tags": [
          "Clinic"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/clinic/skin-archive/{id}": {
      "get": {
        "operationId": "get-clinic-skin-archive-{id}",
        "summary": "Get skin archive scan detail",
        "description": "单条扫描记录；归属不匹配按 404",
        "tags": [
          "Clinic"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRecordDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "operationId": "patch-clinic-skin-archive-{id}",
        "summary": "Update skin archive scan",
        "description": "改趋势（trend）；skin_archive 无 note 列，不支持备注修改",
        "tags": [
          "Clinic"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScanRecordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRecordDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}